+++ src/tiled/documentmanager.h
@@ -101,7 +101,7 @@ public:
     void switchToDocumentAndHandleSimiliarTileset(MapDocument *mapDocument, QPointF viewCenter, qreal scale);
 
     void addDocument(const DocumentPtr &document);
-    void insertDocument(int index, const DocumentPtr &document);
+    int insertDocument(int index, const DocumentPtr &document);
 
     bool isDocumentModified(Document *document) const;